+++ src/benchmark.cpp
@@ -3,6 +3,7 @@
 #include <sstream>
 
 #include "evaluator.hpp"
+#include "evaluator_inc.hpp"
 #include "oeis_sequence.hpp"
 #include "parser.hpp"
 #include "program_util.hpp"
@@ -14,9 +15,16 @@ void Benchmark::all() {
   programs();
 }
 
+std::string fillString(std::string s, size_t n) {
+  while (s.length() < n) {
+    s += " ";
+  }
+  return s;
+}
+
 void Benchmark::operations() {
-  std::cout << "| Operation |  Time   |" << std::endl;
-  std::cout << "|-----------|---------|" << std::endl;
+  std::cout << "| Operation |  Time     |" << std::endl;
+  std::cout << "|-----------|-----------|" << std::endl;
   std::vector<Number> ops(10000);
   int64_t num_digits;
   std::string str;
@@ -58,8 +66,8 @@ void Benchmark::operations() {
     buf.precision(2);
     buf << speed;
     std::cout << "|    "
-              << Operation::Metadata::get(type).name + "    | " + buf.str() +
-                     "µs |"
+              << Operation::Metadata::get(type).name + "    | " +
+                     fillString(buf.str() + "µs", 10) + " |"
               << std::endl;
   }
   std::cout << std::endl;
@@ -67,31 +75,49 @@ void Benchmark::operations() {
 
 void Benchmark::programs() {
   Setup::setProgramsHome("tests/programs");
-  std::cout << "| Sequence | Terms | Time    |" << std::endl;
-  std::cout << "|----------|-------|---------|" << std::endl;
+  std::cout << "| Sequence | Terms  | Reg Eval | Inc Eval |" << std::endl;
+  std::cout << "|----------|--------|----------|----------|" << std::endl;
   program(796, 300);
   program(1041, 300);
   program(1113, 300);
   program(2110, 300);
+  program(57552, 300);
   program(79309, 300);
   program(2193, 400);
+  program(12866, 1000);
   program(45, 2000);
   program(5, 5000);
+  program(130487, 5000);
   program(30, 500000);
   std::cout << std::endl;
 }
 
 void Benchmark::program(size_t id, size_t num_terms) {
   Parser parser;
-  Settings settings;
-  Sequence result;
-  Evaluator evaluator(settings);
-  static const size_t runs = 5;
   const OeisSequence seq(id);
   auto program = parser.parse(seq.getProgramPath());
+  auto speed_reg = programEval(program, false, num_terms);
+  auto speed_inc = programEval(program, true, num_terms);
+  std::cout << "| " << seq.id_str() << "  | "
+            << fillString(std::to_string(num_terms), 6) << " | "
+            << fillString(speed_reg, 8) << " | " << fillString(speed_inc, 8)
+            << " |" << std::endl;
+}
+
+std::string Benchmark::programEval(const Program& p, bool use_inc_eval,
+                                   size_t num_terms) {
+  Settings settings;
+  Interpreter interpreter(settings);
+  IncrementalEvaluator inc_eval(interpreter);
+  if (use_inc_eval && !inc_eval.init(p)) {
+    return "-";
+  }
+  Sequence result;
+  Evaluator evaluator(settings, use_inc_eval);
+  static const size_t runs = 4;
   auto start_time = std::chrono::steady_clock::now();
   for (size_t i = 0; i < runs; i++) {
-    evaluator.eval(program, result, num_terms);
+    evaluator.eval(p, result, num_terms);
     if (result.size() != num_terms) {
       Log::get().error(
           "Unexpected sequence length: " + std::to_string(result.size()), true);
@@ -105,7 +131,6 @@ void Benchmark::program(size_t id, size_t num_terms) {
   std::stringstream buf;
   buf.setf(std::ios::fixed);
   buf.precision(2);
-  buf << speed;
-  std::cout << "| " << seq.id_str() << "  | " << num_terms << "   | "
-            << buf.str() << "s   |" << std::endl;
+  buf << speed << "s";
+  return buf.str();
 }